feat: refactor Map component and introduce hooks for map initialization and zone loading
